lxc-proposed-snapshot: drop support for lxc < 2.1, minor changes.
Support for lxc < 2.1 required using mount-image-callback support
for lxd.  Support for lxd is now removed from mount-image-callback,
so there is not really a way to get this working.  The fallout is
that we cannot use lxd that ships in distro on xenial.  The solution
there is to suggest the snap for lxd.

Other changes here:
 * better debug messages
 * add and use 'find_command' to find lxc-pstart in the same directory
   as lxc-proposed-snapshot.