update-root: re-use code in get-proposed-image and lxc-proposed-snapshot.
This adds a 'update-root' program that applies the updates that
get-proposed-image and lxc-proposed-snapshot use.  It makes those two
programs basically just prepare root, and then execute update-root in
that root.

It gives them similar usage also:
  get-proposed-image --proposed --upgrade=foo xenial [xenial-p.img]
  lxc-proposed-snapshot --proposed --upgrade=foo xenial [xenial-p]

get-proposed-image now does not create the raw image by default, but
can do that with '--raw'.